move pick relays loop to relays service
