+++ lib/services/nostr/nostr_service.dart
@@ -529,34 +529,6 @@ class NostrService {
         .toList();
     log("userFollows: $userFollows");
     log("debug");
-
-    var relaysPicker = RelaysPicker();
-    await relaysPicker.init(pubkeys: userFollows, coverageCount: 2);
-
-    List<RelayAssignment> foundRelays = [];
-    Map<String, int> excludedRelays = {};
-    int now = DateTime.now().millisecondsSinceEpoch ~/ 1000;
-    while (true) {
-      try {
-        var result = relaysPicker.pick(userFollows);
-        var assignment = relaysPicker.getRelayAssignment(result);
-        if (assignment == null) {
-          continue;
-        }
-        foundRelays.add(assignment);
-
-        // exclude already found relays
-        excludedRelays[assignment.relayUrl] = now;
-
-        relaysPicker.setExcludedRelays = excludedRelays;
-      } catch (e) {
-        log("catch: $e");
-        break;
-      }
-    }
-    for (var relay in foundRelays) {
-      log("relay: ${relay.relayUrl}, pubkey: ${relay.pubkeys}");
-    }
-    log("found relays: $foundRelays");
+    relays.selectBestRelays(userFollows);
   }
 }
